+++ 2DGraphicsLU2.WebApi/Repositories/Object2DRepository.cs
@@ -13,53 +13,55 @@ namespace _2DGraphicsLU2.WebApi.Repositories
             this.sqlConnectionString = sqlConnectionString;
         }
 
-        public async Task<Object2D> InsertAsync(Object2D object2D)
+        public async Task<Object2D> InsertAsync(Object2D object2D, string userId)
         {
             using (var sqlConnection = new SqlConnection(sqlConnectionString))
             {
-                var environmentId = await sqlConnection.ExecuteAsync("INSERT INTO [Object2D] (Id, PrefabId, PositionX, PositionY, ScaleX, ScaleY, RotationZ, SortingLayer) VALUES (@Id, @PrefabId, @PositionX, @PositionY, @ScaleX, @ScaleY, @RotationZ, @SortingLayer)", object2D);
+                var environmentId = await sqlConnection.ExecuteAsync("INSERT INTO [Object2D] (Id, PrefabId, PositionX, PositionY, ScaleX, ScaleY, RotationZ, SortingLayer) VALUES (@Id, @PrefabId, @PositionX, @PositionY, @ScaleX, @ScaleY, @RotationZ, @SortingLayer) WHERE userId = @userId", new { object2D, userId});
                 return object2D;
             }
         }
 
-        public async Task<Object2D?> ReadAsync(Guid id)
+        public async Task<Object2D?> ReadAsync(Guid id, string userId)
         {
             using (var sqlConnection = new SqlConnection(sqlConnectionString))
             {
-                return await sqlConnection.QuerySingleOrDefaultAsync<Object2D>("SELECT * FROM [Object2D] WHERE Id = @Id", new { id });
+                return await sqlConnection.QuerySingleOrDefaultAsync<Object2D>("SELECT * FROM [Object2D] WHERE Id = @Id && WHERE userId = @userId", new { id, userId });
             }
         }
 
-        public async Task<IEnumerable<Object2D>> ReadAsync()
+        public async Task<IEnumerable<Object2D>> ReadAsync(string userId)
         {
             using (var sqlConnection = new SqlConnection(sqlConnectionString))
             {
-                return await sqlConnection.QueryAsync<Object2D>("SELECT * FROM [Object2D]");
+                return await sqlConnection.QueryAsync<Object2D>("SELECT * FROM [Object2D] WHERE userId = @userId", new { userId });
             }
         }
 
-        public async Task UpdateAsync(Object2D environment)
+        public async Task UpdateAsync(Object2D environment, string userId)
         {
             using (var sqlConnection = new SqlConnection(sqlConnectionString))
             {
                 await sqlConnection.ExecuteAsync("UPDATE [Object2D] SET " +
                                                  "PrefabId = @PrefabId, " +
-                                                 "PositionX = @PositionX," +
-                                                 "PositionY = @PositionY," +
-                                                 "ScaleX = @ScaleX," +
-                                                 "ScaleY = @ScaleY," +
+                                                 "PositionX = @PositionX, " +
+                                                 "PositionY = @PositionY, " +
+                                                 "ScaleX = @ScaleX, " +
+                                                 "ScaleY = @ScaleY, " +
                                                  "RotationZ = @RotationZ," +
-                                                 "SortingLayer = @SortingLayer"
-                                                 , environment);
+                                                 "SortingLayer = @SortingLayer " +
+                                                 "WHERE Id = @Id " +
+                                                 "&& WHERE userId = @userId"
+                                                 , new { environment, userId });
 
             }
         }
 
-        public async Task DeleteAsync(Guid id)
+        public async Task DeleteAsync(Guid id, string userId)
         {
             using (var sqlConnection = new SqlConnection(sqlConnectionString))
             {
-                await sqlConnection.ExecuteAsync("DELETE FROM [Object2D] WHERE Id = @Id", new { id });
+                await sqlConnection.ExecuteAsync("DELETE FROM [Object2D] WHERE Id = @Id && WHERE userId = @userId", new { id , userId});
             }
         }
